Implicitly cache the properties on first access
This makes the library much easier to use as you don't need to remember what
self.r2_get_xxx_x64() method(s) need to be called before reading public object
properties.

It also hides the implementation of what gets discovered and how from the module
user, which is a good thing because it allows us to change the internals in the
future without breaking API again.